fixing kafka issue. Making sure correct user details is passed on from the gateway to the services.

+++ src/main/java/com/auction/order_service/dto/PurchaseResponse.java
@@ -8,7 +8,7 @@ public record PurchaseResponse(
         String description,
 
         BigDecimal price,
-        double quantity
+        Integer quantity
 
 ) {
